scrapping script/script/requirements.txt
python-dotenv==1.0.1
selenium==4.18.1
webdriver-manager==4.0.1
# Uncomment for dockerfile-gunicorn
# Flask
# gunicorn
google-cloud-firestore